requirements.txt
Flask
numpy
scikit-learn
Pillow
webcolors